pyproject.toml
[project]
name = "docquery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.86.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
    "requests>=2.32.0",
    "streamlit>=1.45.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "json-repair>=0.59.4",
    "jupyter>=1.1.1",
    "markitdown>=0.1.5",
    "matplotlib>=3.10.8",
    "opencv-python>=4.8.0",
    "pdfplumber>=0.11.9",
    "pymupdf-layout>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
